Changes
in the books the examples are using the "using namespace std" and I'm
avoiding to take habit to use std::. Some programmers say which is not
good to use  "using namespace std" because that can be a problem and to
fix complicated. The code was not working before because the part of
getline(cin, str) I put std:: before the getline and the code was not
working to, but I had to put std:: on the cin inside of parenteses to
that I didn't know.